fix windows lint and have complete UT coverage of isIPAffiliated
